# backend/main-fiber/docs/swagger.yaml
basePath: /api
definitions:
...
  api.CreateMembershipReq:
    properties:
      joined_at:
        type: string
      org_path:
        description: '"/club/cpsk"'
        type: string
      position_key:
        description: '"head", "member", "student"'
        type: string
      user_ref:
        description: _id hex OR numeric id OR student_id
        type: string
    type: object
  api.MembershipDoc:
    properties:
      _id:
        type: string
      created_at:
        type: string
      ended_at:
        type: string
      joined_at:
        type: string
      org_path:
        type: string
      position_key:
        type: string
      status:
        description: '"active" | "inactive"'
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  api.PagedUsersResponse:
    properties:
      items:
        description: '[]models.User'
      nextCursor:
        type: string
    type: object
...
  api.UpdateMembershipReq:
    properties:
      ended_at:
        type: string
      position_key:
        type: string
      status:
        description: '"active" | "inactive"'
        type: string
    type: object
...
  models.User:
    properties:
      _id:
        type: string
      advisor_id:
        type: string
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      id:
        description: numeric app id
        type: integer
      lastName:
        type: string
      roles:
        items:
          type: string
        type: array
      student_id:
        type: string
      thaiprefix:
        type: string
      type_person:
        type: string
      updatedAt:
        type: string
    type: object
info:
  contact: {}
  description: This is my API
  title: My API
  version: "1.0"
paths:
...
  /api/memberships:
    get:
      consumes:
      - application/json
      description: List memberships with optional filters
      parameters:
      - description: User reference (_id, id, student_id)
        in: query
        name: user
        type: string
      - description: Organization path
        in: query
        name: org_path
        type: string
      - description: Position key
        in: query
        name: position_key
        type: string
      - description: Membership status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.MembershipDoc'
            type: array
      summary: List memberships
      tags:
      - memberships
    post:
      consumes:
      - application/json
      description: Assign a user to an org unit and position
      parameters:
      - description: Membership info
        in: body
        name: CreateMembershipReq
        required: true
        schema:
          $ref: '#/definitions/api.CreateMembershipReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.MembershipDoc'
      summary: Create membership
      tags:
      - memberships
  /api/memberships/{id}:
    delete:
      description: Delete membership by ID
      parameters:
      - description: Membership ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete membership
      tags:
      - memberships
    get:
      consumes:
      - application/json
      description: Get membership by ID
      parameters:
      - description: Membership ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MembershipDoc'
      summary: Get membership
      tags:
      - memberships
    patch:
      consumes:
      - application/json
      description: Update membership fields (position, status, ended_at)
      parameters:
      - description: Membership ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: UpdateMembershipReq
        required: true
        schema:
          $ref: '#/definitions/api.UpdateMembershipReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MembershipDoc'
      summary: Update membership
      tags:
      - memberships
...
  /users:
    get:
      parameters:
      - description: Search by name/email/student_id
        in: query
        name: q
        type: string
      - description: Page size (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Keyset cursor (base64)
        in: query
        name: cursor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.PagedUsersResponse'
      summary: Get users with search + keyset pagination
      tags:
      - users
    post:
      consumes:
      - application/json
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
      summary: Create a new user
      tags:
      - users
  /users/{id}:
    delete:
      parameters:
      - description: User numeric id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete a user by numeric id
      tags:
      - users
    get:
      parameters:
      - description: User numeric id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Get a single user by numeric app id
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: User numeric id
        in: path
        name: id
        required: true
        type: integer
      - description: User partial or full
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Update an existing user by numeric id
      tags:
      - users
swagger: "2.0"